Update API version to 2024-05-01 across resources
Updated the API version from "2022-08-01" to "2024-05-01" in multiple test files and templates related to Azure API Management resources. This ensures alignment with the latest API version, incorporating new features and improvements. Changes were made in various test classes, including `DeploymentTemplateTests`, `FunctionAppMixedWithLogicAppTests`, and `LogicAppAddedAsOperationTests`, as well as in `template.json`, `templateParameters.json`, and the `MockResourceCollector` class. The `Constants.cs` file was also updated to reflect the correctversion.